+++ cpp/open3d/core/Tensor.cpp
@@ -33,10 +33,10 @@
 #include "open3d/core/Device.h"
 #include "open3d/core/Dispatch.h"
 #include "open3d/core/Dtype.h"
-#include "open3d/core/Kernel/Kernel.h"
 #include "open3d/core/ShapeUtil.h"
 #include "open3d/core/SizeVector.h"
 #include "open3d/core/TensorKey.h"
+#include "open3d/core/kernel/Kernel.h"
 #include "open3d/utility/Console.h"
 
 namespace open3d {
