+++ cpp/open3d/core/kernel/BinaryEWCPU.cpp
@@ -24,14 +24,14 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/core/Kernel/BinaryEW.h"
+#include "open3d/core/kernel/BinaryEW.h"
 
 #include "open3d/core/Dispatch.h"
 #include "open3d/core/Dtype.h"
-#include "open3d/core/Kernel/CPULauncher.h"
 #include "open3d/core/MemoryManager.h"
 #include "open3d/core/SizeVector.h"
 #include "open3d/core/Tensor.h"
+#include "open3d/core/kernel/CPULauncher.h"
 #include "open3d/utility/Console.h"
 
 namespace open3d {
