+++ cpp/open3d/core/kernel/IndexGetSet.cpp
@@ -24,13 +24,13 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/core/Kernel/IndexGetSet.h"
+#include "open3d/core/kernel/IndexGetSet.h"
 
 #include "open3d/core/Dtype.h"
-#include "open3d/core/Kernel/UnaryEW.h"
 #include "open3d/core/MemoryManager.h"
 #include "open3d/core/SizeVector.h"
 #include "open3d/core/Tensor.h"
+#include "open3d/core/kernel/UnaryEW.h"
 #include "open3d/utility/Console.h"
 
 namespace open3d {
