+++ cpp/open3d/core/kernel/NonZeroCPU.cpp
@@ -24,7 +24,7 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/core/Kernel/NonZero.h"
+#include "open3d/core/kernel/NonZero.h"
 
 #include "open3d/core/Indexer.h"
 #include "open3d/utility/Console.h"
