+++ cpp/open3d/core/kernel/NonZeroCUDA.cu
@@ -24,7 +24,7 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/core/Kernel/NonZero.h"
+#include "open3d/core/kernel/NonZero.h"
 
 #include <thrust/device_vector.h>
 #include <thrust/execution_policy.h>
