+++ cpp/open3d/core/kernel/ReductionCPU.cpp
@@ -24,7 +24,7 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/core/Kernel/Reduction.h"
+#include "open3d/core/kernel/Reduction.h"
 
 #include <limits>
