+++ cpp/open3d/core/kernel/UnaryEWCUDA.cu
@@ -24,13 +24,13 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/core/Kernel/UnaryEW.h"
+#include "open3d/core/kernel/UnaryEW.h"
 
 #include "open3d/core/CUDAState.cuh"
 #include "open3d/core/CUDAUtils.h"
 #include "open3d/core/Dispatch.h"
-#include "open3d/core/Kernel/CUDALauncher.cuh"
 #include "open3d/core/Tensor.h"
+#include "open3d/core/kernel/CUDALauncher.cuh"
 
 namespace open3d {
 namespace core {
