+++ cpp/tests/core/Tensor.cpp
@@ -29,10 +29,10 @@
 
 #include "open3d/core/AdvancedIndexing.h"
 #include "open3d/core/Dtype.h"
-#include "open3d/core/Kernel/Kernel.h"
 #include "open3d/core/MemoryManager.h"
 #include "open3d/core/SizeVector.h"
 #include "open3d/core/Tensor.h"
+#include "open3d/core/kernel/Kernel.h"
 #include "open3d/utility/Helper.h"
 
 #include "tests/UnitTest.h"
